backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
django-environ==0.11.2
django-cors-headers==4.3.1
python-dateutil==2.9.0
gunicorn==21.2.0
